+++ bin/dice_tractogram_filter.py
@@ -13,6 +13,7 @@ parser.add_argument("--minweight", type=float, help="Keep streamlines with weigh
 parser.add_argument("--maxweight", type=float, help="Keep streamlines with weight <= this value")
 parser.add_argument("--weights_in",  help="Text file with the input streamline weights")
 parser.add_argument("--weights_out", help="Text file for the output streamline weights")
+parser.add_argument("--random", type=float, default=1.0, help="Randomly discard streamlines: 0=discard all, 1=keep all")
 parser.add_argument("--verbose", "-v", action="store_true", help="Print information messages")
 parser.add_argument("--force", "-f",   action="store_true", help="Force overwriting of the output")
 options = parser.parse_args()
@@ -27,6 +28,7 @@ filter(
     options.maxweight,
     options.weights_in,
     options.weights_out,
+    options.random,
     options.verbose,
     options.force
 )